bugfix: Fix the segmentation fault error when closing the application by commenting out the delete operation in the GraphNode class
Looked throu the code to get a better understanding what is going on in the chatbot implementation and annotate the individual
elements with their functionallity. Also the bug of the segmentation fault while closing the chatbot application, introduced in the
lecture, is fixed. The error was that the destructor of ChatBot was called twice, since the ChatLogic destructor calls the delete
operation on the ChatBot instance, as well as the GraphNode destructor. The correct point of deletion is from the current point
of view the ChatLogic class, since there are many GraphNode instances and the destructor of ChatLogic is called before the destructor
of the first GraphNode instance is called.

No issue addressed