feat: Implement the rule of five for the ChatBot class
Implement the rule of five for the ChatBot class in chatbot.h and chatbot.cpp.
The exclusive ownership policy is implemented since the ChatBot instance
moves as a unique element throu the answer graph. We do not want to create two
ChatBot instances that are walking throu the graph (at least for this assignment.
Maybe it is a cool idea to create a chatbot that answers multiple users questions
with one program instance, like a server that can be called by user clients over
the network)

No issue addressed