feat: Make the GraphEdge elements owned by the GraphNode elements (fulfull task 4)
In order to fulfill task 4, the GraphEdge elements were made exclusivly owned by the
GraphNode class. The idea behind this is, that the ChatLogic class creates the GraphNode
and GraphEdge elements and by incooperating move semantics, the ownership is transferred
from the ChatLogic class to the GraphNode class (which all a are members of ChatLogic).
Thereby the ownership is transferred in a divide and conquer manner.

No issue addressed