feat: Make the ChatBot instance an exclusivly managed ressource and move it throu the graph using the rule of five
Make the ChatBot instance an exclusivly managed ressource by the GraphNode elements using unique_ptr and
move semantics by incoperating the rule of five that was implemented in task 2 within chatbot.cpp and
chatbot.h. The code does compile but there is a segmentation fault on runtime that needs to be fixed before the
assignment is handed in.

No issue addressed